refactor: change paths to components animations
